requirements.txt
Flask==3.0.3
google-api-python-client==2.141.0
google-auth-oauthlib==1.2.1
google-auth==2.34.0
PyPDF2==3.0.1
python-dotenv==1.0.1
google-genai==0.6.0
